+++ docs/source/News.rst
@@ -5,6 +5,16 @@
 This document describes the releases of :mod:`repoze.what.plugins.sql`.
 
 
+.. _repoze.what-sql-1.0b2:
+
+:mod:`repoze.what.plugins.sql` 1.0b2 (*unreleased*)
+===================================================
+
+* Introduced the plugin :class:`FriendlyRedirectingFormPlugin
+  <repoze.what.plugins.quickstart.FriendlyRedirectingFormPlugin>` and used by
+  default in :func:`repoze.what.plugins.quickstart.setup_sql_auth`.
+
+
 .. _repoze.what-sql-1.0b1:
 
 :mod:`repoze.what.plugins.sql` 1.0b1 (2009-01-27)